# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ProcurementFlow API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Department {
  id: number;
  name: string;
}

export interface InventoryItem {
  id: number;
  name: string;
  sku: string;
  stockLevel: number;
  unitPrice: number;
  category: string;
  lowStockThreshold: number;
  createdAt?: string;
}

export interface InventoryItemInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  sku: string;
  /** @minimum 0 */
  stockLevel: number;
  /** @minimum 0 */
  unitPrice: number;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  lowStockThreshold: number;
}

export interface InventoryItemUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  sku?: string;
  /** @minimum 0 */
  stockLevel?: number;
  /** @minimum 0 */
  unitPrice?: number;
  category?: string;
  /** @minimum 0 */
  lowStockThreshold?: number;
}

export type RequestStatus = typeof RequestStatus[keyof typeof RequestStatus];


export const RequestStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  fulfilled: 'fulfilled',
} as const;

export interface Request {
  id: number;
  employeeEmail: string;
  departmentId: number;
  itemId: number;
  quantity: number;
  status: RequestStatus;
  createdAt: string;
  /** @nullable */
  departmentName?: string | null;
  /** @nullable */
  itemName?: string | null;
  /** @nullable */
  itemSku?: string | null;
}

export interface RequestInput {
  /** @minLength 1 */
  employeeEmail: string;
  departmentId: number;
  itemId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface Transaction {
  id: number;
  requestId: number;
  itemId: number;
  departmentId: number;
  quantityDeducted: number;
  timestamp: string;
  /** @nullable */
  itemName?: string | null;
  /** @nullable */
  departmentName?: string | null;
}

export interface DashboardSummary {
  totalRequests: number;
  pendingRequests: number;
  approvedRequests: number;
  rejectedRequests: number;
  totalInventoryItems: number;
  lowStockCount: number;
  totalInventoryValue: number;
}

export interface DepartmentUsage {
  departmentId: number;
  departmentName: string;
  totalQuantity: number;
  totalRequests: number;
}

export interface WebhookPayload {
  employeeEmail: string;
  departmentName: string;
  itemName: string;
  quantity: number;
}

export interface WebhookResult {
  received: boolean;
  /** @nullable */
  message?: string | null;
}

export type ListInventoryItemsParams = {
category?: string;
lowStock?: boolean;
};

export type ListRequestsParams = {
status?: ListRequestsStatus;
departmentId?: number;
};

export type ListRequestsStatus = typeof ListRequestsStatus[keyof typeof ListRequestsStatus];


export const ListRequestsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  fulfilled: 'fulfilled',
} as const;

export type ListTransactionsParams = {
departmentId?: number;
itemId?: number;
};