# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ProcurementFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalRequests: number;
  pendingRequests: number;
  approvedRequests: number;
  rejectedRequests: number;
  totalInventoryItems: number;
  lowStockCount: number;
  totalInventoryValue: number;
}