# lib/api-zod/src/generated/types/departmentUsage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ProcurementFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DepartmentUsage {
  departmentId: number;
  departmentName: string;
  totalQuantity: number;
  totalRequests: number;
}