# lib/api-zod/src/generated/types/inventoryItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ProcurementFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItem {
  id: number;
  name: string;
  sku: string;
  stockLevel: number;
  unitPrice: number;
  category: string;
  lowStockThreshold: number;
  createdAt?: string;
}