# lib/api-zod/src/generated/types/inventoryItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ProcurementFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItemInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  sku: string;
  /** @minimum 0 */
  stockLevel: number;
  /** @minimum 0 */
  unitPrice: number;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  lowStockThreshold: number;
}