# lib/api-zod/src/generated/types/requestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ProcurementFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RequestInput {
  /** @minLength 1 */
  employeeEmail: string;
  departmentId: number;
  itemId: number;
  /** @minimum 1 */
  quantity: number;
}