# lib/api-zod/src/generated/types/webhookResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ProcurementFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WebhookResult {
  received: boolean;
  /** @nullable */
  message?: string | null;
}